Add files via upload
Contains additional code for scraping data from the 1997 webpage. There are still a few minor kinks to iron out, which is why I haven't appended the actual data yet, but I figured it might be helpful for anyone who is currently pulling their hair out trying to navigate a sea of <p> tags and nonsensical formatting exceptions. I think a similar general approach may work on some of the later pages as well.