Add files via upload
The script I used to scrape most of the years and the data for those years. I think my script for 2001-2009 can be tweaked to work with the earlier years (98-00); it probably won't work with 2010. Let me know if anyone finds any errors or modifications that need to be made.