Add files via upload
Code used to split out authors. I created a new table which lists every author for every paper (each row contains one paper-author-affiliation combination, so if a paper had 3 authors, it would have 3 consecutive records in the table). We should probably make a separate "authors" table at some point (in database terms, "papers" and "authors" are entities in an M-to-M relationship, and this table is the extra table created to capture the relationship between the two in a relational setting...right now we don't have an "authors" table).

There are still some issues that need to be sorted out - suffixes that have commas in them, affiliations with nested parentheses (apparently regex can't really handle them, so I left them out), a few records with bizarre formatting, etc. However, these are fairly minor; I would say this is 90-95% of the final goal. Please review and let me know if anyone finds any errors.